+++ tests/test_train.py
@@ -8,13 +8,13 @@ import torch.optim as optim
 from torch.utils.data import DataLoader
 
 from pytorch_ner.dataset import NERCollator, NERDataset
+from pytorch_ner.logger import get_logger
 from pytorch_ner.prepare_data import (
     get_label2idx,
     get_token2idx,
     prepare_conll_data_format,
 )
 from pytorch_ner.train import train_loop, validate_epoch
-from pytorch_ner.utils import get_logger
 from tests.test_nn_modules.test_architecture import model_bilstm as model
 
 logger = get_logger()
